+++ docs/changes.md
@@ -1,4 +1,7 @@
 # 最新动态
+2025/07/26
+  * 完善darray/dlist
+
 2025/07/24
   * 删除多余的path_extname_is原型声明(感谢兆坤发现问题)
   * 修复api comment